+++ R/svm_linear.R
@@ -17,9 +17,12 @@
 #'  functions. If parameters need to be modified, `update()` can be used
 #'  in lieu of recreating the object from scratch.
 #'
-#' @param mode A single character string for the type of model.
+#' @param mode A single character string for the prediction outcome mode.
 #'  Possible values for this model are "unknown", "regression", or
 #'  "classification".
+#' @param engine A single character string specifying what computational engine
+#'  to use for fitting. Possible engines are listed below. The default for this
+#'  model is `"LiblineaR"`.
 #' @param cost A positive number for the cost of predicting a sample within
 #'  or on the wrong side of the margin
 #' @param margin A positive number for the epsilon in the SVM insensitive
@@ -45,7 +48,7 @@
 #' @export
 
 svm_linear <-
-  function(mode = "unknown",
+  function(mode = "unknown", engine = "LiblineaR",
            cost = NULL, margin = NULL) {
 
     args <- list(
@@ -59,7 +62,7 @@ svm_linear <-
       eng_args = NULL,
       mode = mode,
       method = NULL,
-      engine = NULL
+      engine = engine
     )
   }